+++ src/commands/ext-update.ts
@@ -11,7 +11,6 @@ import {
   logPrefix,
   getSourceOrigin,
   SourceOrigin,
-  confirm,
   diagnoseAndFixProject,
   isLocalPath,
 } from "../extensions/extensionsHelper";
@@ -21,7 +20,7 @@ import * as refs from "../extensions/refs";
 import { getProjectId } from "../projectUtils";
 import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
-import * as experiments from "../experiments";
+import { confirm } from "../prompt";
 import * as manifest from "../extensions/manifest";
 import { Options } from "../options";
 import * as askUserForEventsConfig from "../extensions/askUserForEventsConfig";
@@ -35,9 +34,7 @@ marked.setOptions({
  */
 export const command = new Command("ext:update <extensionInstanceId> [updateSource]")
   .description(
-    experiments.isEnabled("extdev")
-      ? "update an existing extension instance to the latest version or from a local or URL source"
-      : "update an existing extension instance to the latest version"
+    "update an existing extension instance to the latest version, or to a specific version if provided"
   )
   .before(requirePermissions, [
     "firebaseextensions.instances.update",
@@ -46,19 +43,10 @@ export const command = new Command("ext:update <extensionInstanceId> [updateSour
   .before(ensureExtensionsApiEnabled)
   .before(checkMinRequiredVersion, "extMinVersion")
   .before(diagnoseAndFixProject)
-  .option("--local", "deprecated")
   .withForce()
   .action(async (instanceId: string, updateSource: string, options: Options) => {
     const projectId = getProjectId(options);
     const config = manifest.loadConfig(options);
-
-    if (options.local) {
-      utils.logLabeledWarning(
-        logPrefix,
-        "As of firebase-tools@11.0.0, the `--local` flag is no longer required, as it is the default behavior."
-      );
-    }
-
     const oldRefOrPath = manifest.getInstanceTarget(instanceId, config);
     if (isLocalPath(oldRefOrPath)) {
       throw new FirebaseError(
@@ -104,11 +92,11 @@ export const command = new Command("ext:update <extensionInstanceId> [updateSour
     );
 
     if (
-      !(await confirm({
+      !await confirm({
         nonInteractive: options.nonInteractive,
         force: options.force,
         default: false,
-      }))
+      })
     ) {
       utils.logLabeledBullet(logPrefix, "Update aborted.");
       return;
