+++ src/extensions/extensionsHelper.ts
@@ -30,7 +30,7 @@ import {
 import { ExtensionSource, ExtensionVersion, Param } from "./types";
 import * as refs from "./refs";
 import { getLocalExtensionSpec } from "./localHelper";
-import { promptOnce } from "../prompt";
+import { promptOnce, confirm } from "../prompt";
 import { logger } from "../logger";
 import { envOverride } from "../utils";
 import { getLocalChangelog } from "./change-log";
@@ -789,28 +789,6 @@ export function getSourceOrigin(sourceOrVersion: string): SourceOrigin {
       sourceOrVersion
     )}'. Check to make sure the source is correct, and then please try again.`
   );
-}
-
-/**
- * Confirm if the user wants to continue
- */
-export async function confirm(args: {
-  nonInteractive?: boolean;
-  force?: boolean;
-  default?: boolean;
-}): Promise<boolean> {
-  if (!args.nonInteractive && !args.force) {
-    const message = `Do you wish to continue?`;
-    return await promptOnce({
-      type: "confirm",
-      message,
-      default: args.default,
-    });
-  } else if (args.nonInteractive && !args.force) {
-    throw new FirebaseError("Pass the --force flag to use this command in non-interactive mode");
-  } else {
-    return true;
-  }
 }
 
 export async function diagnoseAndFixProject(options: any): Promise<void> {
